requirements.txt
Pillow==9.3.0
progress==1.5
pytest
pytest-cov
codecov
coverage
flask
numpy
replicate
requests
